// src/lib/moetruyen/model/getV1CommentsChaptersById200DataItemRepliesItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Moetruyen Public API
 * Unofficial read-only REST API for [MoeTruyen](https://moetruyen.net/).

Full features will be added in the future (chắc thế)

Github: [TNTKien/moetruyen-public-api](https://github.com/TNTKien/moetruyen-public-api), [dex593/web1 (MoeTruyen)](https://github.com/dex593/web1)

NOTE:
- All API endpoints have a global rate limit of 7 requests per second per IP.
- To avoid future issues, include the Origin: https://suicaodex.com or https://moetruyen.net headers when making API requests.
 * OpenAPI spec version: 0.1.0
 */
import type { GetV1CommentsChaptersById200DataItemRepliesItemAuthor } from "./getV1CommentsChaptersById200DataItemRepliesItemAuthor";

export type GetV1CommentsChaptersById200DataItemRepliesItem = {
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  id: number;
  /** @minLength 1 */
  content: string;
  createdAt: string | null;
  /** @minLength 1 */
  commentPath: string;
  author: GetV1CommentsChaptersById200DataItemRepliesItemAuthor;
};